requirements.txt
llama-index==0.12.1
llama-index-embeddings-huggingface==0.4.0
python-docx==1.1.2
fitz==0.0.1.dev2
PyMuPDF==1.24.14
easyocr==1.7.2
chainlit==1.1.304
langchain-huggingface==0.1.2
langchain-google-genai==2.0.5
langchain-community==0.3.7
langchain-chroma==0.1.4
torch==2.5.1


